+++ jest.config.ts
@@ -3,7 +3,7 @@ import type { Config } from 'jest'
 const config: Config = {
     preset: 'jest-preset-angular',
     setupFilesAfterEnv: ['<rootDir>/jest.setup.ts'],
-    testPathIgnorePatterns: ['<rootDir>/node_modules/', '<rootDir>/dist/', '<rootDir>/cypress/'],
+    testPathIgnorePatterns: ['<rootDir>/node_modules/', '<rootDir>/dist/', '<rootDir>/cypress/', '<rootDir>/.cache/'],
     coverageReporters: ['html', 'text-summary', 'cobertura', 'lcov'],
     coverageDirectory: 'coverage',
     moduleNameMapper: {